requirements.txt
Appium-Python-Client
pytest
selenium